+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/School.kt
@@ -5,7 +5,7 @@ import com.sexysisters.tojserverv2.domain.school.exception.SchoolException
 import com.sexysisters.tojserverv2.domain.school.type.Division
 import com.sexysisters.tojserverv2.domain.school.type.Kind
 import com.sexysisters.tojserverv2.domain.student.domain.Student
-import com.sexysisters.tojserverv2.domain.teacher.Teacher
+import com.sexysisters.tojserverv2.domain.teacher.domain.Teacher
 import com.sexysisters.tojserverv2.domain.wiki.Wiki
 import javax.persistence.*
