+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/teacher/TeacherEntityMapper.kt
@@ -1,5 +1,6 @@
 package com.sexysisters.tojserverv2.domain.teacher
 
+import com.sexysisters.tojserverv2.domain.teacher.domain.*
 import org.springframework.stereotype.Component
 
 @Component
