+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/teacher/TeacherStore.kt
@@ -1,5 +1,7 @@
 package com.sexysisters.tojserverv2.domain.teacher
 
+import com.sexysisters.tojserverv2.domain.teacher.domain.Teacher
+
 interface TeacherStore {
     fun store(teacher: Teacher): Teacher
 }
