+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/teacher/domain/Bio.kt
@@ -1,20 +1,20 @@
-package com.sexysisters.tojserverv2.domain.teacher
+package com.sexysisters.tojserverv2.domain.teacher.domain
 
 import com.sexysisters.tojserverv2.domain.teacher.exception.TeacherException
 import javax.persistence.Column
 import javax.persistence.Embeddable
 
 @Embeddable
-class Name private constructor (
-    @Column(name = "name", unique = true)
+class Bio private constructor(
+    @Column(name = "bio")
     val value: String
 ) {
     companion object {
-        private val REGEX = Regex("^{1,100}$")
+        private val REGEX = Regex("^{1,500}$")
 
-        fun of(value: String): Name {
+        fun of(value: String): Bio {
             validate(value)
-            return Name(value)
+            return Bio(value)
         }
 
         private fun validate(value: String) {
