+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/teacher/domain/Teacher.kt
@@ -1,4 +1,4 @@
-package com.sexysisters.tojserverv2.domain.teacher
+package com.sexysisters.tojserverv2.domain.teacher.domain
 
 import com.sexysisters.tojserverv2.domain.BaseTimeEntity
 import com.sexysisters.tojserverv2.domain.school.School
