+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/teacher/TeacherRepository.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.teacher
 
-import com.sexysisters.tojserverv2.domain.teacher.Teacher
+import com.sexysisters.tojserverv2.domain.teacher.domain.Teacher
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface TeacherRepository : JpaRepository<Teacher, Long> {
